Bugfix a wrong behavior on the calculate function
Update the value of the previousValue variable to empty on lines 93,
97, 103 & 108 on the calculate function.

When the user used the equals to trigger the calculate function and
then tried to use an operator to perform a chain calculation using
the last result, the previousValue kept the same value.

As a result, the handleOperator function triggered the calculate
function instantly (line 55).

To work correctly, the result of the operation after using equals should
be our new previousValue and then wait for the new currentValue to
perform another operation with equals or an operator.